Custom components

+++ src/app/pages/crud-storage/crud-storage.page.ts
@@ -1,48 +1,79 @@
-import { Component, ViewChild } from '@angular/core';
-import { Platform, ToastController, IonList } from '@ionic/angular';
+import { Component, ViewChild, OnInit } from '@angular/core';
+import { Platform, ToastController, IonList, LoadingController } from '@ionic/angular';
 import { StorageService, Item } from 'src/app/services/storage.service';
+import { AuthService } from 'src/app/services/auth.service';
 
 @Component({
   selector: 'app-crud-storage',
   templateUrl: './crud-storage.page.html',
   styleUrls: ['./crud-storage.page.scss'],
 })
-export class CrudStoragePage {
+export class CrudStoragePage implements OnInit {
 
   items: Item[] = [];
 
   newItem: Item = <Item>{};
   @ViewChild('mylist', { static: false }) mylist: IonList;
 
-  constructor(private storageService: StorageService, private plt: Platform, private toastController: ToastController) {
+  constructor(private loadCtrl: LoadingController, private auth: AuthService, private storageService: StorageService, private plt: Platform, private toastController: ToastController) {
     this.plt.ready().then(() => {
       this.loadItems();
     });
   }
+  data: any;
+  ngOnInit() {
+    // const loading = await this.loadCtrl.create({
+    //   message: 'Loading data ...',
+    // });
+    // await loading.present();
+    // this.loadItems();
+    // this.data = '1';
+    // console.log(this.data)
+    // await loading.dismiss();
+  }
+  signOut() {
+    this.auth.signOut();
+  }
 
   // CREATE
   addItem() {
-    this.newItem.modified = Date.now();
-    this.newItem.id = Date.now();
-    this.storageService.addItem(this.newItem).then(item => {
-      this.newItem = <Item>{};
-      this.showToast('Item added!')
-      this.loadItems(); // Or add it to the array directly
-    });
+    if (this.newItem.id) {
+      this.showToast('this is not a new item.')
+    } else {
+      this.newItem.modified = Date.now();
+      this.newItem.id = Date.now();
+      this.storageService.addItem(this.newItem).then(item => {
+        this.newItem = <Item>{};
+        this.showToast('Item added!')
+        this.loadItems(); // Or add it to the array directly
+      });
+    }
+
   }
 
   clearItem(item: Item) {
     if (!item.id) {
       this.showToast('no data to clear');
+      this.newItem = <Item>{};
     } else {
+      this.showToast('your data\'s clear');
       this.newItem = <Item>{};
     }
   }
+
   // READ
-  loadItems() {
-    this.storageService.getItems().then(items => {
-      this.items = items;
+  async loadItems() {
+    const loading = await this.loadCtrl.create({
+      message: 'Loading data ...',
     });
+    await loading.present();
+
+    this.items = await this.storageService.getItems()
+    if (this.items != null) {
+      sortByKey(this.items, 'modified')
+    }
+    this.data = '1';
+    await loading.dismiss();
   }
 
   // READ ONE ITEM
@@ -83,12 +114,18 @@ export class CrudStoragePage {
 
   // DELETE
   deleteItem(item: Item) {
-    this.storageService.deleteItem(item.id).then(item => {
+    if (!item.id) {
       this.newItem = <Item>{};
-      this.showToast('Item removed!');
-      this.mylist.closeSlidingItems(); // Fix or sliding is stuck afterwards
-      this.loadItems(); // Or splice it from the array directly
-    });
+      this.showToast('no data to remove');
+    } else {
+      this.storageService.deleteItem(item.id).then(item => {
+        this.newItem = <Item>{};
+        this.showToast('Item removed!');
+        this.mylist.closeSlidingItems(); // Fix or sliding is stuck afterwards
+        this.loadItems(); // Or splice it from the array directly
+      });
+    }
+
   }
 
   // Helper
@@ -101,3 +138,16 @@ export class CrudStoragePage {
   }
 
 }
+function sortByKey(array: Item[], key: string) {
+  return array.sort(function (a: { [x: string]: any; }, b: { [x: string]: any; }) {
+    var x = a[key];
+    var y = b[key];
+    return ((x > y) ? -1 : ((x > y) ? 1 : 0));
+  });
+}
+  // function sortByKey(array, key) {
+  //   return array.sort(function(a, b) {
+  //       var x = a[key]; var y = b[key];
+  //       return ((x < y) ? -1 : ((x > y) ? 1 : 0));
+  //   });
+  //   }
